Removed glutton for-in loops and useless function calls
Tried to transform recursives functions (_scan & isValid) into iterative functions but it made no difference.

About 15% gain with check() and 25% with isValid() : 
 - 1400ms instead of 1650ms for 20000 calls of check()
 - 1400ms instead of 1850ms for 50000 calls of isValid()

Use custom type does'nt change anything