@jlesceau 2015-04-17
In a case the type isn't valid or the data don't match the type, we want the scanning function to check a small number of values before getting to the error. Wouldn't be interesting to scan first all scalar types and then go through deep objects or arrays ?
It could also result in a loss of perfs when sorting the keys, especially when the type is correct..