@jlesceau 2015-04-20
New way to test if exclusive or optionnal ( charAt is much faster than match ).
Skip the 'unexpected key' test in a object when the expected types are here and the number of keys is the same.
New way to test if a object has a given key ( http://jsperf.com/hasownproperty-vs-in-vs-undefined/12 )