@jlesceau 2015-04-22
| Test check() | Modified | Original |
| :------------- | :-------------: | :-------------: |
| Deep object |||
| noError | 1664ms | 1583ms |
| errorBeginning | 152ms | 1104ms |
| errorEnd | 1561ms | 2917ms |
| unexpectedKey |  1683ms | 2430ms |
| ---- | ---- | ---- |
| Large object of strings |||
| noError | 2118ms | 4922ms |
| errorBeginning | 190ms | 1171ms |
| errorEnd | 2086ms | 5969ms |
| unexpectedKey | 2180ms | 6013ms |
| ---- | ---- | ---- |
| Large object of functions |||
| noError | 2383ms | 5124ms |
| errorBeginning | 188ms | 1187ms |
| errorEnd | 2269ms | 6253ms |
| unexpectedKey | 2389ms | 6032ms |
| ---- | ---- | ---- |
| Array of strings |||
| noError | 2104ms | 4498ms |
| errorBeginning | 114ms | 1098ms |
| errorEnd | 2188ms | 5627ms |
| ---- | ---- | ---- |
| Array of customs |||
| noError | 151ms | 1750ms |
| errorBeginning | 258ms | 2083ms |
| errorEnd | 138ms | 1755ms |