@jlesceau 2015-04-22
| Test isValid() | Modified | Original |
| :---- | :----: | :----: |
| Deep object |||
| noError | 960ms | 1073ms |
| errorBeginning | 123ms | 126ms |
| errorEnd | 946ms | 1039ms |
| ---- | ---- | ---- |
| Large object of strings |||
| noError | 2119ms | 3287ms |
| errorBeginning | 159ms | 197ms |
| errorEnd | 2088ms | 3201ms |
| ---- | ---- | ---- |
| Large object of functions |||
| noError | 2086ms | 2998ms |
| errorBeginning | 164ms | 189ms |
| errorEnd | 2038ms | 3731ms |
| ---- | ---- | ---- |
| Array of strings |||
| noError | 89ms | 210ms |
| error | 109ms | 238ms |
| ---- | ---- | ---- |
| Array of customs|||
| noError | 95ms | 141ms |