@jacomyal 2015-04-23
Here are some notes about the object returned by `_scan` when the value matches the type:
  - I don't like the `"optional"`key. Maybe `"nully"` might be more appropriate?
  - I think the `"type"` of the value should be returned as well.