@jacomyal 2015-04-23
More precisely, here are some notes that might make `_scan` more consistent:
  - It always returns an object, with the keys **type**, **expected** and **value**.
  - If the type does not match, then the error message is returned under the key **error**.
  - If the type is `undefined` or `null` and the `?` operator was used, then the **nully** flag has the value `true`.

What do you think about that?