@jlesceau 2015-04-23
| Test isValid() | Modified | Original | Ratio |
| :------------- | :-------------: | :-------------: | :-------------: |
| **Deep object** ||||
| noError | 922ms | 975ms | -0.05 |
| errorBeginning | 114ms | 114ms | 0 |
| errorEnd | 912ms | 979ms | -0.07 |
| **Large object of strings** ||||
| noError | 1909ms | 2542ms | -0.25 |
| errorBeginning | 160ms | 176ms | -0.09 |
| errorEnd | 1946ms | 2793ms | -0.3 |
| **Large object of functions** ||||
| noError | 1946ms | 2546ms | -0.24 |
| errorBeginning | 157ms | 173ms | -0.09 |
| errorEnd | 1840ms | 2468ms | -0.25 |
| **Array of strings** ||||
| noError | 83ms | 120ms | -0.31 |
| error | 102ms | 128ms | -0.2 |
| **Array of customs** ||||
| noError | 92ms | 127ms | -0.28 |