@jlesceau 2015-04-23
| Test check() | Modified | Original | Ratio |
| :------------- | :-------------: | :-------------: | :-------------: |
| **Deep object** ||||
| noError | 1647ms | 1573ms | +0.05 |
| errorBeginning | 145ms | 1054ms | -0.86 |
| errorEnd | 1650ms | 2822ms | -0.42 |
| unexpectedKey | 1720ms | 2329ms | -0.26 |
| **Large object of strings** ||||
| noError | 2344ms | 4756ms | -0.51 |
| errorBeginning | 201ms | 1139ms | -0.82 |
| errorEnd | 2275ms | 5871ms | -0.61 |
| unexpectedKey | 2336ms | 5789ms | -0.6 |
| **Large object of functions** ||||
| noError | 2397ms | 4968ms | -0.52 |
| errorBeginning | 204ms | 1138ms | -0.82 |
| errorEnd | 2163ms | 5790ms | -0.63 |
| unexpectedKey | 2246ms | 5794ms | -0.61 |
| **Array of strings** ||||
| noError | 2028ms | 4379ms | -0.54 |
| errorBeginning | 116ms | 1092ms | -0.89 |
| errorEnd | 2176ms | 5487ms | -0.6 |
| **Array of customs** ||||
| noError | 9167ms | 14600ms | -0.37 |
| errorBeginning | 302ms | 1987ms | -0.85 |
| errorEnd | 9321ms | 16219ms | -0.43 |