@jlesceau 2015-04-23
| Test check() | Modified | Original | Ratio |
| :------------- | :-------------: | :-------------: | :-------------: |
| **Deep object** ||||
| noError | 1763ms | 1595ms | +0.11 |
| errorBeginning | 145ms | 1060ms | -0.86 |
| errorEnd | 1551ms | 2832ms | -0.45 |
| unexpectedKey | 1672ms | 2362ms | -0.29 |
| **Large object of strings** ||||
| noError | 2275ms | 4641ms | -0.51 |
| errorBeginning | 195ms | 1121ms | -0.83 |
| errorEnd | 2230ms | 5794ms | -0.62 |
| unexpectedKey | 2352ms | 5818ms | -0.6 |
| **Large object of functions** ||||
| noError | 2455ms | 5044ms | -0.51 |
| errorBeginning | 202ms | 1151ms | -0.82 |
| errorEnd | 2235ms | 6410ms | -0.65 |
| unexpectedKey | 2343ms | 5971ms | -0.61 |
| **Array of strings** ||||
| noError | 2099ms | 4521ms | -0.54 |
| errorBeginning | 116ms | 1130ms | -0.9 |
| errorEnd | 2187ms | 5726ms | -0.62 |
| **Array of customs** ||||
| noError | 9051ms | 14633ms | -0.38 |
| errorBeginning | 253ms | 2051ms | -0.88 |
| errorEnd | 9060ms | 16733ms | -0.46 |